requirements.txt
pyyaml
tqdm
numpy
scipy
scikit-learn
torch
transformers
fastapi[standard]
PyMuPDF
python-docx
extract_msg
sqlalchemy